+++ packages/zefyr/lib/src/widgets/selection.dart
@@ -10,8 +10,7 @@ import 'package:zefyr/util.dart';
 
 import 'controller.dart';
 import 'editable_box.dart';
-import 'editable_text.dart';
-import 'editor.dart';
+import 'scope.dart';
 
 RenderEditableBox _getEditableBox(HitTestResult result) {
   for (var entry in result.path) {
@@ -73,14 +72,14 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
   }
 
   void showToolbar() {
-    final editable = ZefyrEditableText.of(context);
-    assert(editable != null);
+    final scope = ZefyrScope.of(context);
+    assert(scope != null);
     final toolbarOpacity = _toolbarController.view;
     _toolbar = new OverlayEntry(
       builder: (context) => new FadeTransition(
             opacity: toolbarOpacity,
             child: new _SelectionToolbar(
-              editable: editable,
+              scope: scope,
               controls: widget.controls,
               delegate: this,
             ),
@@ -117,7 +116,7 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    final editor = ZefyrEditor.of(context);
+    final editor = ZefyrScope.of(context);
     if (_editor != editor) {
       _editor?.removeListener(_handleChange);
       _editor = editor;
@@ -174,7 +173,7 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
   OverlayEntry _toolbar;
   AnimationController _toolbarController;
 
-  ZefyrEditorScope _editor;
+  ZefyrScope _editor;
   TextSelection _selection;
   FocusOwner _focusOwner;
 
@@ -190,9 +189,9 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
     if (!mounted) {
       return;
     }
-    final editor = ZefyrEditor.of(context);
-    final selection = editor.selection;
-    final focusOwner = editor.focusOwner;
+
+    final selection = _editor.selection;
+    final focusOwner = _editor.focusOwner;
     setState(() {
       if (focusOwner != FocusOwner.editor) {
         hideToolbar();
@@ -233,8 +232,8 @@ class _ZefyrSelectionOverlayState extends State<ZefyrSelectionOverlay>
 
     RenderEditableProxyBox box = _getEditableBox(result);
     if (box == null) {
-      final editable = ZefyrEditableText.of(context);
-      box = editable.renderContext.closestBoxForGlobalPoint(globalPoint);
+      final scope = ZefyrScope.of(context);
+      box = scope.renderContext.closestBoxForGlobalPoint(globalPoint);
     }
     if (box == null) return null;
 
@@ -327,8 +326,8 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    final editable = ZefyrEditableText.of(context);
-    _selection = editable.selection;
+    final scope = ZefyrScope.of(context);
+    _selection = scope.selection;
   }
 
   //
@@ -337,15 +336,14 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
 
   @override
   Widget build(BuildContext context) {
-    final editor = ZefyrEditor.of(context);
-    final editable = ZefyrEditableText.of(context);
+    final scope = ZefyrScope.of(context);
     if (selection == null ||
         selection.isCollapsed ||
         widget.controls == null ||
-        editor.focusOwner != FocusOwner.editor) {
+        scope.focusOwner != FocusOwner.editor) {
       return new Container();
     }
-    final block = editable.renderContext.boxForTextOffset(documentOffset);
+    final block = scope.renderContext.boxForTextOffset(documentOffset);
     final position = getPosition(block);
     Widget handle;
     if (position == null) {
@@ -395,9 +393,8 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
   void _handleDragUpdate(DragUpdateDetails details) {
     _dragPosition += details.delta;
     final globalPoint = _dragPosition;
-    final editor = ZefyrEditor.of(context);
-    final editable = ZefyrEditableText.of(context);
-    final paragraph = editable.renderContext.boxForGlobalPoint(globalPoint);
+    final scope = ZefyrScope.of(context);
+    final paragraph = scope.renderContext.boxForGlobalPoint(globalPoint);
     if (paragraph == null) {
       return;
     }
@@ -414,7 +411,7 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
     }
 
     if (newSelection != _selection) {
-      editor.updateSelection(newSelection, source: ChangeSource.local);
+      scope.updateSelection(newSelection, source: ChangeSource.local);
     }
   }
 }
@@ -422,12 +419,12 @@ class _SelectionHandleDriverState extends State<SelectionHandleDriver> {
 class _SelectionToolbar extends StatefulWidget {
   const _SelectionToolbar({
     Key key,
-    @required this.editable,
+    @required this.scope,
     @required this.controls,
     @required this.delegate,
   }) : super(key: key);
 
-  final ZefyrEditableTextScope editable;
+  final ZefyrScope scope;
   final TextSelectionControls controls;
   final TextSelectionDelegate delegate;
 
@@ -436,7 +433,7 @@ class _SelectionToolbar extends StatefulWidget {
 }
 
 class _SelectionToolbarState extends State<_SelectionToolbar> {
-  ZefyrEditableTextScope get editable => widget.editable;
+  ZefyrScope get editable => widget.scope;
   TextSelection get selection => widget.delegate.textEditingValue.selection;
 
   @override