+++ packages/zefyr/lib/zefyr.dart
@@ -22,6 +22,7 @@ export 'src/widgets/list.dart';
 export 'src/widgets/paragraph.dart';
 export 'src/widgets/quote.dart';
 export 'src/widgets/scaffold.dart';
+export 'src/widgets/scope.dart' hide ZefyrScopeAccess;
 export 'src/widgets/selection.dart' hide SelectionHandleDriver;
 export 'src/widgets/theme.dart';
 export 'src/widgets/toolbar.dart';