+++ packages/zefyr/CHANGELOG.md
@@ -1,3 +1,16 @@
+## Unreleased
+
+* Added non-scrollable `ZefyrView` widget which allows previewing Notus documents inside
+  layouts using their own scrollables like ListView.
+* Breaking change: renamed `EditableRichText` to `ZefyrRichText`. User code is unlikely to be
+  affected unless you've extended Zefyr with custom implementations of block widgets.
+* Breaking change: renamed `RenderEditableParagraph` to `RenderZefyrParagraph`. User code is
+  unlikely to be affected unless you've extended Zefyr with custom implementations of block widgets.
+* Added `ZefyrScope` class - replaces previously used scope objects `ZefyrEditableTextScope` and
+  `ZefyrEditorScope`. Unified all shared resources under one class.
+* Breaking change: removed `ZefyrEditor.of` and `ZefyrEditableText.of` static methods.
+  Use `ZefyrScope.of` instead.
+
 ## 0.3.0
 
 This version introduces new widget `ZefyrScaffold` which allows embedding Zefyr in custom